move context_holder to job scope;rename get_data_loader_by_sharding -> get_sharding_data_loader
