+++ client/starwhale/api/_impl/dataset/loader.py
@@ -157,7 +157,7 @@ def get_data_loader(
     return _cls(dataset_uri, start, end, logger or _logger)
 
 
-def get_data_loader_by_sharding(
+def get_sharding_data_loader(
     dataset_uri: str = "",
     sharding_index: int = 0,
     sharding_num: int = 1,
